requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langchain>=0.0.300
httpx>=0.24.1
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic-ai-slim==0.0.46
pydantic-core==2.33.1
pydantic-graph==0.0.46
pydantic-settings~=2.9.1
langchain-core~=0.1.53